Separate ImportWorkflow tests
